Fixed the problem with the number of questions
The number of questions was hardcoded as 10, this caused a problem when the number of questions available was lesser than 10.
Due to this, the result page didn't show up when needed. Fixed the issue by deciding the number of questions to be asked
dynamically. The max number of quesions is still 10, but if 10 questions aren't available, all of the questions are asked.
Result calculation was updated to suit the number of questions.